refactor: extract capability::probe (loopback discovery, health probing, Tier cache)
Fifth step of the capability.rs module split, stacked on the guard.rs
extraction. Pure lift-and-shift: spelunk_state_dir, read_server_port_file,
the TIER cache, get_tier, probe_tier_fresh, the probe timeouts/port
constant, probe/probe_url/parse_health, and current_uid (plus tests)
move into capability/probe.rs unchanged. capability.rs re-exports the
external-facing pieces.

Depends on diagnostics.rs, state.rs, and tier.rs (all already
extracted); lands fifth. What remains of capability.rs after this step
is just the module doc plus mod declarations and re-exports, i.e.
already equivalent to the eventual capability/mod.rs content.
